Re-enable auto-freeze/at-risk cron jobs; fix telegram_id int32 overflow
Incident follow-up now that the real root cause (348e623) is fixed and
confirmed live:

- Restore streak-freeze-auto-apply and streak-at-risk-push to their
  original hourly schedule, and streak-reminder to hourly-with-local-
  hour-filter. These were paused as a diagnostic guess during the
  incident (connection-pool contention) that turned out to be wrong —
  they were never the actual cause, safe to bring back.

- migration 087: user.telegram_id, book_purchase.telegram_id,
  book_rating.telegram_id, user_quiz_completion.telegram_id were all
  32-bit INTEGER, overflowing on modern Telegram IDs (e.g.
  7984840136 > INTEGER max ~2.1B). Found via a real traceback
  (profiles.py get_my_rating) that surfaced in the same log window as
  the incident but is unrelated to it. book_read_progress.telegram_id
  (a newer table) already used BIGINT correctly — same fix applied to
  the three/four legacy tables that didn't. ALTER COLUMN TYPE to
  BIGINT is non-destructive, strictly widens the column.

+++ backend/app/models/models.py
@@ -312,7 +312,7 @@ class User(Base):
     __tablename__ = "user"
     
     id = Column(Integer, primary_key=True, index=True)
-    telegram_id = Column(Integer, unique=True, index=True)
+    telegram_id = Column(BigInteger, unique=True, index=True)  # 087_fix_telegram_id_integer_overflow
     username = Column(String(255), unique=True, index=True, nullable=True)
     first_name = Column(String(255), nullable=True)
     last_name = Column(String(255), nullable=True)
@@ -489,7 +489,7 @@ class BookPurchase(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     book_id = Column(Integer, ForeignKey("book.id"), index=True)
-    telegram_id = Column(Integer, index=True)                   # buyer
+    telegram_id = Column(BigInteger, index=True)                # buyer — 087_fix_telegram_id_integer_overflow
     provider = Column(String(30), index=True)                   # click | payme
     provider_transaction_id = Column(String(255), nullable=True) # external tx id
     order_id = Column(String(100), unique=True, index=True)     # our internal ref
@@ -506,7 +506,7 @@ class BookRating(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     book_id = Column(Integer, ForeignKey("book.id"), index=True)
-    telegram_id = Column(Integer, index=True)
+    telegram_id = Column(BigInteger, index=True)  # 087_fix_telegram_id_integer_overflow
     rating = Column(Integer)  # 1-5
     review = Column(Text, nullable=True)
     created_at = Column(DateTime, default=datetime.utcnow)
@@ -547,7 +547,7 @@ class UserQuizCompletion(Base):
     id = Column(Integer, primary_key=True, index=True)
     user_id = Column(Integer, ForeignKey("user.id"), index=True, nullable=True)
     quiz_id = Column(Integer, ForeignKey("quiz.id"), index=True)
-    telegram_id = Column(Integer, index=True)  # denormalize for speed
+    telegram_id = Column(BigInteger, index=True)  # denormalize for speed — 087_fix_telegram_id_integer_overflow
     score = Column(Integer)
     total = Column(Integer)
     percentage = Column(Float)
